src/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
prometheus-client==0.20.0
pydantic==2.7.1